+++ R/build_clusters.R
@@ -34,8 +34,6 @@
 #'
 #' # investigate variable relations
 #' rel <- var.relations(
-#'   x = data.frame(),
-#'   create.forest = FALSE,
 #'   forest = list(trees = trees.surr, allvariables = allvariables),
 #'   variables = allvariables,
 #'   candidates = allvariables,