+++ NEWS.md
@@ -1,5 +1,4 @@
-<!-- News Style-guide: https://style.tidyverse.org/news.html -->
-# RFSurrogates 0.4.0
+# RFSurrogates (development version)
 
 # RFSurrogates 0.3.3