+++ NEWS.md
@@ -1,6 +1,4 @@
-# RFSurrogates 0.3.4
-
-<!-- News Style-guide: https://style.tidyverse.org/news.html -->
+# RFSurrogates (development version)
 
 * `var.select.smd()`, `var.select.md()`, `var.relations()`, `var.relations.mfi()`: Made several improvements to developer experience:
   * `create.forest` now defaults to `is.null(forest)`, so it will automatically be `TRUE` if no forest is provided, and `FALSE` otherwise.